core: Fix tag_iterator failure on empty tags.
This code used to be `< 0`, but was changed to `<= 0` in commit d40ddfe.

Thinking more about this, it's a bad idea to abort on 0 length tags,
since it's completely valid for some tags (such as SSID) to be empty. It
doesn't make sense to check if `tag_len` is below 0, because it's
unsigned.

So, remove the check completely.